ai-service/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
openai
pinecone-client
python-dotenv
